BUGFIX: Do not depend on structure of parent class
This is a backwards compatible fix for changes in the Flow PdoBackend.
The backend in this class depended on the ``initializeObject`` method
in the parent but that was technically not necessary and with an
inject method for the configuration we make this independent from
it's parent class and any changes of it.

Fixes: #3